Remove text_decoder callback in ctcdecode-numpy as requested
