+++ demos/speech_recognition_demo/python/utils/ctc_decoder_seq_pipeline.py
@@ -3,11 +3,28 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 from copy import deepcopy
+import codecs
 
 from ctcdecode_numpy import SeqCtcLmDecoder
 
-from utils.alphabet import load_alphabet, CtcdecoderAlphabet
 from utils.pipelines import BlockedSeqPipelineStage
+
+
+def load_alphabet(filename):
+    characters = []
+    with codecs.open(filename, 'r', 'utf-8') as f:
+        for line in f:
+            line = line.rstrip('\r\n')
+            if line == '':  # empty line ends the alphabet
+                break
+            if line[0] == '#':  # comment
+                continue
+            if line.startswith('\\s'):  # "\s" for space as the first character
+                line = ' ' + line[2:]
+            elif line[0] == '\\':  # escaping, to enter "#" or "\" as the first character
+                line = line[1:]
+            characters.append(line)
+    return characters
 
 
 class CtcDecoderSeqPipelineStage(BlockedSeqPipelineStage):
@@ -16,16 +33,13 @@ class CtcDecoderSeqPipelineStage(BlockedSeqPipelineStage):
         self.p = deepcopy(profile)
         self.online = online
 
-        alphabet = self.p['alphabet']
         if isinstance(self.p['alphabet'], str):
             alphabet = load_alphabet(self.p['alphabet'])  # shall not include <blank> token
         else:
             alphabet = self.p['alphabet']  # list-like
 
-        alphabet_decoder = CtcdecoderAlphabet(alphabet)
         self._decoder = SeqCtcLmDecoder(alphabet, beam_width, max_candidates=max_candidates,
-            scorer_lm_fname=lm, alpha=self.p['alpha'], beta=self.p['beta'],
-            text_decoder=lambda symbols: alphabet_decoder.decode(symbols))
+            scorer_lm_fname=lm, alpha=self.p['alpha'], beta=self.p['beta'])
 
         super().__init__(
             block_len=1, context_len=0,
